Move startup code to app.js
